#34 refactor add moreinterestedmedicine package
