+++ feature/interestedmedicine/src/main/res/navigation/interestedmedicine_nav.xml
@@ -11,7 +11,7 @@
 
     <fragment
         android:id="@+id/moreInterestedMedicineFragment"
-        android:name="com.android.mediproject.feature.interestedmedicine.MoreInterestedMedicineFragment"
+        android:name="com.android.mediproject.feature.interestedmedicine.moreinterestedmedicine.MoreInterestedMedicineFragment"
         android:label="MoreInterestedMedicineFragment" >
         <deepLink app:uri="medilens://main/moreInterestedMedicine_nav" />
     </fragment>
